Rename some safe txs
