+++ src/utils/safe.ts
@@ -1,8 +1,4 @@
-import { CV_V2, CV_V3 } from 'constants/cv'
-import { useLoadV2V3Contract } from 'hooks/v2v3/LoadV2V3Contract'
 import { GnosisSafe, SafeTransactionType } from 'models/safe'
-import { V2V3ContractName } from 'models/v2v3/contracts'
-import { CV2V3 } from 'models/v2v3/cv'
 
 // e.g. [ {nonce: 69}, {nonce: 45}, {nonce: 69}] returns [69, 45]
 export function getUniqueNonces(
@@ -43,25 +39,4 @@ export function isSafeSigner({
   return safe.owners.some(
     owner => owner.toLowerCase() === address.toLowerCase(),
   )
-}
-
-export function getTransactionVersion(
-  transaction: SafeTransactionType,
-): CV2V3 | undefined {
-  const V2JBController = useLoadV2V3Contract({
-    cv: CV_V2,
-    contractName: V2V3ContractName.JBController,
-  })
-  const V3JBController = useLoadV2V3Contract({
-    cv: CV_V3,
-    contractName: V2V3ContractName.JBController,
-  })
-
-  if (transaction.to === V2JBController?.address) {
-    return CV_V2
-  }
-
-  if (transaction.to === V3JBController?.address) {
-    return CV_V3
-  }
 }
